requirements/dev.txt
-r base.txt
python-dotenv==1.0.1
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
django-debug-toolbar==4.4.6
ipython==8.31.0
